requirements.txt
groq>=0.4.0
openai>=1.0.0
anthropic>=0.18.0
supabase>=2.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
bandit>=1.7.0
safety>=3.0.0
pip-audit>=1.0.0
pyyaml>=6.0
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
numpy>=1.24.0
scikit-learn>=1.3.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
